created pagination component. working on issues with making api call after state updates
