+++ client/src/App.js
@@ -1,6 +1,5 @@
 import './App.css';
-import axios from 'axios';
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import Home from './views/Home';
 import Login from './views/Login';
